python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai-whisper==20240930
python-multipart==0.0.17
pydub==0.25.1
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
